Revert changes to `extract-metadata.py` script which are now in #8
